requirements.txt
dask[complete]>=2023.11
graphviz
boto3
botocore
scikit-learn
numpy>=1.25.*
pandas>=2.0.*
scipy>=1.11.*
requests>=2.31.*
PyYAML
pytest
coverage
pytest-cov
pylint
flake8